refactor(extract): begin per-language extractor split (#1212)
